backend/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
sqlparse==0.5.4
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
requests==2.31.0
beautifulsoup4==4.12.3
unidecode==1.3.8


